pyproject.toml
[project]
name = "api-openai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "openai>=2.16.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
